+++ backend/internal/endpoint/user.go
@@ -65,6 +65,11 @@ func Register(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{"result": "invalid institution ID"})
 		return
 	}
+	ins := database.InstitutionInfoByID(uint(insID))
+	if ins.ID == 0 {
+		c.JSON(http.StatusBadRequest, gin.H{"result": "no institution found"})
+		return
+	}
 	user.InstitutionID = uint(insID)
 
 	err = database.CreateUser(user)
@@ -74,57 +79,6 @@ func Register(c *gin.Context) {
 	}
 
 	c.JSON(http.StatusOK, user)
-}
-
-// Login godoc
-//	@Summary		login
-//	@Description	get a session
-//	@Tags			Session
-//	@Accept			json
-//	@Produce		json
-//	@Param			username	formData	string				true	"user's name"
-//	@Param			password	formData	string				true	"password"
-//	@Success		200			{object}	response.Response	"success | has loginned"
-//	@Failure		401			{object}	response.Response	"wrong username or password"
-//	@Router			/session [post]
-func Login(c *gin.Context) {
-	username := c.PostForm("username")
-	password := c.PostForm("password")
-
-	if pkg.IsAuthenticated(c) {
-		c.JSON(http.StatusOK, gin.H{"result": "has loginned"})
-		return
-	}
-
-	var user database.User
-	user = database.FindByUserName(username)
-
-	// No user found
-	if user.ID == 0 {
-		c.JSON(http.StatusUnauthorized, gin.H{"result": "wrong username or password"})
-		return
-	}
-
-	// Wrong password
-	if err := pkg.Compare(user.Password, password); err != nil {
-		c.JSON(http.StatusUnauthorized, gin.H{"result": "wrong username or password"})
-		return
-	}
-
-	pkg.SaveAuthSession(c, user.ID, user.Name)
-	c.JSON(http.StatusOK, gin.H{"result": "success"})
-}
-
-// Logout godoc
-//	@Summary		logout
-//	@Description	delete the session
-//	@Tags			Session
-//	@Produce		json
-//	@Success		200	{object}	response.Response	"success"
-//	@Router			/session [delete]
-func Logout(c *gin.Context) {
-	pkg.ClearAuthSession(c)
-	c.JSON(http.StatusOK, gin.H{"result": "success"})
 }
 
 // ModifyInfo godoc
